About the SUBDOMAIN EXTRAVAGANZA
I wrote this software as part of a DESEC SECURITY task. This is my first functional and usefull program ever!
The purpose of this is to acquire knowledge and know-how to become a RED HAT cybersecurity professional.

This tool parses HTML domains, informed by the user, and convert it automatically into IP Addresses.
It is a mapping tool.

Even though I was fully able to develop it by myself, with any help of a guiding and breathing alive human, I still do not fully understand it's possibilities of use. I shall continue my studies and utilize it as much as I possibly can, in order to learn and develop more skills.

Whilst this is an open-source program, I would like to affirm that I use it for legal purposes only. And so should you! Any crime committed by you is your sole responsibility. Don`t do it! Otherwise, help yourself with this tool and learn. Become a pro and help others ethically!